requirements.txt
alembic==1.12.1
blinker==1.6.3
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
flask==3.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
flask-sqlalchemy==3.1.1
greenlet==3.0.1
gunicorn==21.2.0
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
packaging==23.2
pluggy==1.3.0
protobuf==4.21.12
PyJWT==2.8.0
pytest==7.4.3
pytest-flask==1.3.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
tomli==2.0.1
typing-extensions==4.8.0
werkzeug==3.0.1
zipp==3.17.0
